[CH012] SRP implementation WIP
