+++ src/modules/user.module.ts
@@ -4,9 +4,10 @@ import { UserController } from 'src/controllers/user.controller';
 import { UserService } from 'src/services/user.service';
 import { AuthService } from 'src/services/auth.service';
 import { Users } from 'src/entities/user.entity';
+import { Keywords } from 'src/entities/keywords.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Users])],
+  imports: [TypeOrmModule.forFeature([Users, Keywords])],
   providers: [UserService, AuthService],
   controllers: [UserController],
   exports: [UserService],
