+++ src/processors/scrape.processor.ts
@@ -1,20 +1,26 @@
 import { Process, Processor } from '@nestjs/bull';
 import { Job } from 'bull';
-import { DashboardService } from 'src/services/dashboard.service';
+import { GetGooglePageService } from 'src/services/scraper/getGooglePage.service';
+import { ScrapeGooglePage } from 'src/services/scraper/scrapeGooglePage.service';
+import { SaveScrapedData } from 'src/services/scraper/saveScrapedData.service';
 
 @Processor('scrape-queue')
 export class ScrapeProcessor {
-  constructor(private dashboardService: DashboardService) {}
+  constructor(
+    private getGooglePageService: GetGooglePageService,
+    private scrapeGooglePage: ScrapeGooglePage,
+    private saveScrapedData: SaveScrapedData,
+  ) {}
 
   @Process('initate_worker_process')
   async initateWorkerProcess(job: Job) {
-    const htmlData = await this.dashboardService.searchGoogleAndReturnHTML(
+    const htmlData = await this.getGooglePageService.getGooglePage(
       job.data.keyword,
     );
 
-    const totalResults = await this.dashboardService.getTotalResults(htmlData);
+    const totalResults = await this.scrapeGooglePage.getTotalResults(htmlData);
 
-    await this.dashboardService.saveDataToKeywordsonDB(
+    await this.saveScrapedData.saveDataToKeywordsonDB(
       job.data.id,
       totalResults,
       htmlData,
